backend/requirements.txt
# SearchX Backend - Python Requirements

# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.25

# Image processing
Pillow>=10.2.0
PyMuPDF>=1.23.8

# OCR
pytesseract>=0.3.10

# Document processing
python-docx>=1.1.0

# ML and embeddings
sentence-transformers>=2.3.1
torch>=2.1.2
faiss-cpu>=1.7.4
numpy>=1.24.3

# Utilities
aiofiles>=23.2.1

# PDF Generation
reportlab>=4.0.0

# Hybrid Search
rank-bm25>=0.2.0
